LayoutTests/media/modern-media-controls/icon-service/icon-service-expected.txt

Testing the iconService singleton.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Checking path to images is computed according to traits
PASS iconService.urlForIconNameAndLayoutTraits(Icons.Pause, LayoutTraits.macOS).includes("macOS/pause") is true
PASS iconService.urlForIconNameAndLayoutTraits(Icons.Pause, LayoutTraits.iOS).includes("iOS/pause") is true
PASS iconService.urlForIconNameAndLayoutTraits(Icons.Pause, LayoutTraits.macOS | LayoutTraits.Fullscreen).includes("macOS/pause-fullscreen") is true

Checking requested images are cached
PASS iconService.imageForIconNameAndLayoutTraits(Icons.Pause, LayoutTraits.macOS) === iconService.imageForIconNameAndLayoutTraits(Icons.Pause, LayoutTraits.macOS) is true
PASS successfullyParsed is true

TEST COMPLETE
